Move some utility functions into their own module
