+++ foilpy/LL_functions.py
@@ -479,48 +479,3 @@ def u_gamma_remove_nan_inf(u_gamma):
 def update_elmt_length(nodes1, nodes2):
     length = np.sqrt(np.sum((nodes2-nodes1) ** 2, axis=1))
     return length
-
-
-def rotation_matrix(w, angle, deg=True):
-    # w is the axis to rotate about, size (3,)
-    # angle is the angle to rotate by in radian
-    if deg == True:
-        angle = angle * np.pi / 180
-    c = np.cos(angle)
-    s = np.sin(angle)
-    R = np.array([[w[0] ** 2 * (1 - c) + c, w[0] * w[1] * (1 - c) - w[2] * s, w[0] * w[2] * (1 - c) + w[1] * s, 0],
-                  [w[1] * w[0] * (1 - c) + w[2] * s, w[1] ** 2 * (1 - c) + c, w[1] * w[2] * (1 - c) - w[0] * s, 0],
-                  [w[2] * w[0] * (1 - c) - w[1] * s, w[2] * w[1] * (1 - c) + w[0] * s, w[2] ** 2 * (1 - c) + c, 0],
-                  [0, 0, 0, 1]])
-
-    return R
-
-
-def apply_rotation(R, vec, dim):
-    # R is rotation matrix size (4,4)
-    # vec is set of vectors (m,3), or (3,m)
-    # dim is the dimension that the vector is in: 0 or 1
-
-    if dim == -1:
-        vec = np.vstack((vec.reshape(3,-1), np.ones((1, 1))))
-        vec_rot = np.dot(R, vec)
-        vec_rot = vec_rot[0:3, :].reshape(-1)
-    elif dim == 0:
-        vec = np.vstack((vec, np.ones((1, vec.shape[1]))))
-        vec_rot = np.dot(R, vec)
-        vec_rot = vec_rot[0:3, :]
-    elif dim == 1:
-        vec = np.hstack((vec, np.ones((vec.shape[0], 1)))).T
-        vec_rot = np.dot(R, vec).T
-        vec_rot = vec_rot[:, 0:3]
-
-    return vec_rot
-
-
-def translation_matrix(t):
-    T = np.array([[1, 0, 0, t[0]],
-                  [0, 1, 0, t[1]],
-                  [0, 0, 1, t[2]],
-                  [0, 0, 0, 1]])
-    return T
-
